Moving release code over to Cloudsmith.

+++ src/build/shell/release.sh
@@ -17,8 +17,8 @@
 # License along with Geoprism(tm).  If not, see <http://www.gnu.org/licenses/>.
 #
 
-#if curl -f -s --head "http://nexus.terraframe.com/service/local/artifact/maven/redirect?r=allrepos&g=net.geoprism&a=geoprism-web&p=war&v=$GP_VERSION" | head -n 1 | grep "HTTP/1.[01] [23].." > /dev/null; then
-#   echo "The release version $GP_VERSION has already been deployed! Please ensure you are releasing the correct version of geoprism."
+#if curl -f -s --head "http://nexus.terraframe.com/service/local/artifact/maven/redirect?r=allrepos&g=net.geoprism&a=geoprism-web&p=war&v=$VERSION" | head -n 1 | grep "HTTP/1.[01] [23].." > /dev/null; then
+#   echo "The release version $VERSION has already been deployed! Please ensure you are releasing the correct version of geoprism."
 #   exit 1
 #fi
 
@@ -26,8 +26,10 @@ git config --global user.name "$GIT_TF_BUILDER_USERNAME"
 git config --global user.email builder@terraframe.com
 
 cd $WORKSPACE/geoprism
+BRANCH=${GIT_BRANCH#*/}
 
-git checkout dev
+# Jenkins locks us onto a specific commit for some reason. We need to be on that branch to be able to do commits.
+git checkout $BRANCH
 git pull
 
 mvn license:format
@@ -35,24 +37,26 @@ git add -A
 git diff-index --quiet HEAD || git commit -m 'License headers'
 git push
 
-git checkout master
-git merge dev
-git push
+if [ "$BRANCH" == "dev" ]; then
+  git checkout master
+  git merge dev
+  git push
+fi
 
+mvn release:prepare -B -Dtag=$VERSION \
+                 -DreleaseVersion=$VERSION \
+                 -DdevelopmentVersion=$NEXT
 
-mvn release:prepare -B -Dtag=$GP_VERSION \
-                 -DreleaseVersion=$GP_VERSION \
-                 -DdevelopmentVersion=$GP_NEXT
-                 
 mvn release:perform -B -Darguments="-Dmaven.javadoc.skip=true -Dmaven.site.skip=true"
 
-
-cd ..
-rm -rf rwdev
-mkdir rwdev
-cd rwdev
-git clone -b master git@github.com:terraframe/geoprism.git
-cd geoprism
-git checkout dev
-git merge master
-git push
+if [ "$BRANCH" == "dev" ]; then
+  cd ..
+  rm -rf rwdev
+  mkdir rwdev
+  cd rwdev
+  git clone -b master git@github.com:terraframe/geoprism.git
+  cd geoprism
+  git checkout dev
+  git merge master
+  git push
+fi
